Bugfix. skip_children must be False if the module is mandatory.
Explanation of the bug:
-----------------------
In the CMS, modules can be marked as 'optional'. If it is optional, the module and
its risks can be skipped.

When a user is assessing a tool, he chooses to answer the module or not. This
choice is stored in the 'skip_children' column in the 'tree' table. The tree
table contains all nodes (modules, profile questions, risks etc.) per
session_id.

So, skip_children represents user input. It tells us whether the user has
decided to answer an optional module or not.

If we now in the CMS, change this module to make it mandatory and publish this
change, then the user's SQL session is updated with the new information, but
because skip_children is deemed to be user input, it is retained.

So, in the user's session, a mandatory module now still has skip_children=True
and if you navigate to that module in the client, it will skip all the children
risks, and also not show them in the identification report.