Bugfix. treeChanges is not aware of changes that change modules from optional to mandatory.
This is necessary because the session data needs to be updated to ensure that
skip_children=False for the (newly) mandatory module.

Without this fix, skip_children might be True, which means that the mandatory
module gets skipped.

Implementation of the fix:
--------------------------
- Add 'optional' key to the dicts created by getSurveyTree and getSessionTree
- In treeChanges, check for the case where an module is now mandatory but the
session tree node is still have skip_children=True