Remove type="number" from datepicker inputs.
There is a bug in patters which causes an event listener on an input of
type="number" to be called repeatedly (and somewhere, on each iteration the
value of that input is incremented by one).

We need listeners on those inputs in order to update the restricted date ranges
when one date changes.