Register a custom membrane user manager.
The default membrane user manager doesn't give any IStatusMessage messages.

We therefore had to add a hack/workaround in the template to inform the user
that their login failed.

However, now that we have configurable lockout and password policies, this
approach doesn't cut it anymore as we don't have the necessary info on the
template to know what message to give the user.

For example, did the login fail or is the user's account locked?
If lockout is enabled, how do we inform the user of this, or tell them how many
attempts they still have left?

These problems go away if we use the IStatusMessage adapter to flash messages.