Revert "Register a custom membrane user manager."
This reverts commit 559e8d2c8aa3df4af78e7ca65351073b88891437.

I recognize now a problem with using IStatusMessages in the membrane user
manager.

PlonePAS plugins don't have a holistic view of the system and cannot know
whether there are any additional similar plugins that will be called.

So the membrane user manager can only check if the credentials are for a
membrane user, but cannot say with certainty that this user is not registered
at all in the system, since there might be other plugins that will authenticate
the user.

That's why IStatusMessages aren't used, because you don't know whether the
user's login has failed completely, only whether it has failed in this
particular plugin (one of potentially many).

I've added the previous commit to a new branch: custom-user-manager