Fix survey settings ambiguity
Since version 19 was released, it was not possible anymore to display the training slides
without the questions.

This is due to the fact that the setting "enable_web_training" was used for two
purposes: to enable the slides and to enable the questions.

Historically it was only responsible to show the questions.

This patch restores the possibility to display the training slides without the questions by decoupling the two settings.
The setting "enable_web_training" enables the slides, while the new "enable_test_questions" setting controls the visibility of the questions.

The upgrade step sets "enable_test_questions" to True if "enable_web_training" is True and we have training questions.

Refs. https://github.com/syslabcom/scrum/issues/3887