Test for auto csrf protection with SQLAlchemy.
This is NOT FOR MERGE!

Refs https://github.com/syslabcom/scrum/issues/3393

* Updates the `@@start` view: if you add `?write=1`, we do a write on GET.
  Sample url: http://localhost:9660/Plone/client/eu/covid-19/covid-19-revision/++session++79/@@start?write=1
  This changes the title of the session, adding a date and time.
  Extra logging here shows that this makes the SQL session "dirty", so it knows there are changes.
  This is good.
* Prints debugging info in our custom `EuphorieProtectTransform` to see if we detect changes in SQLAlchemy objects.
  Result: no, the SQL session here is *not* dirty, so we think there are no changes, so the write on GET succeeds.
  You can test this in `psql` with `select title from session where id=79;`
  This is bad.
* Adds a very simple `@@protect-test` page which tries the same thing.  Call this on a country.
  Sample url: http://localhost:9660/Plone/client/eu/@@protect-test
  Here it works as we want, and we get redirected to the `@@confirm-action` page.
  This is good.

I added a few commented-out lines in the `start` view that make a simple change to the Plone site description.
Then we *do* get redirected.  So the standard csrf protection still works in this view.

So: theoretically our `EuphorieProtectTransform` transform works, but in practice it does not.

One theory I have is that it would be better in the whole code base to always instantiate an sql Session.
So don't do `Session.query(...)` but always `Session().query(...)`.
But when I do the "wrong" thing in the protect-test view, it still works.  So as theories go, it seems not so good. ;-)